extracted preprocessing
